+++ tests/Haya.Core.Tests/Code.fs
@@ -21,7 +21,7 @@ module Code =
         match attr with
         | :? ResponsibilityAttribute as r -> sprintf "[Haya.ResponsibilityAttribute(Description = \"%s\")]" r.Description
         | :? MetaAttribute as m ->
-            sprintf "[assembly:Haya.MetaAttribute(AppName = \"%s\", Description = \"%s\", Team = \"%s\", System = \"%s\", Repository = \"%s\")]" m.AppName m.Description m.Team m.System m.Repository
+            sprintf "[assembly:Haya.MetaAttribute(AppName = \"%s\", Description = \"%s\", Team = \"%s\", System = \"%s\", Repository = \"%s\")]" m.AppName m.Description m.Owner m.System m.Repository
         | :? CollaboratorAttribute as c ->
             sprintf "[Haya.CollaboratorAttribute(Direction = %s, Protocol = \"%s\", DataDescription = \"%s\", Description = \"%s\", AppName = \"%s\", System = \"%s\", Relationship = %s, Repository = \"%s\")]" (c.Direction.ToString()) c.Protocol c.DataDescription c.Description c.AppName c.System (c.Relationship.ToString()) c.Repository
         | _ -> attr.ToString()